[Symfony2] Used Symfony2 parameters in the legacy application
The `web/config/parameters.php` file now uses the values found
in the `app/config/parameters.yml` file.
With this, there's only one place to put the configuration.

Also the 'Project configuration' section in the README has been
removed. A next commit will make the creation and filling of
the `app/config/parameters.yml` automatic at the installation
of the project.